correct matpred is_triu
